requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
codewars-test-teey==0.1.0
distlib==0.3.8
filelock==3.13.4
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.1
platformdirs==4.2.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.6
virtualenv==20.25.3
yarl==1.9.2
